Extract `extract_result`
