+++ testtools/tests/__init__.py
@@ -14,6 +14,7 @@ def test_suite():
         test_compat,
         test_content,
         test_content_type,
+        test_deferred,
         test_deferredmatchers,
         test_deferredruntest,
         test_distutilscmd,
@@ -35,6 +36,7 @@ def test_suite():
         test_compat,
         test_content,
         test_content_type,
+        test_deferred,
         test_deferredmatchers,
         test_deferredruntest,
         test_distutilscmd,
